% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDatabase.R
\name{CheckDatabase}
\alias{CheckDatabase}
\title{Return number of databases reporting RNA editing event}
\usage{
CheckDatabase(...)
}
\arguments{
\item{...}{one or more vectors giving database annotations for each RNA
editing event}
}
\value{
A numeric vector with database counts
}
\description{
\code{CheckDatabase} returns the number of databases reporting an RNA editing
event
}
\examples{
\dontrun{
db <- CheckDatabase(
RNAEdDataQCed$Annotation$RADAR_gid,
RNAEdDataQCed$Annotation$DARNED_gid,
RNAEdDataQCed$Annotation$REDIportal_gid
)
table(db)
}
}
