% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineBaseCounts.R
\name{CombineBaseCounts}
\alias{CombineBaseCounts}
\title{Combine BaseCounts}
\usage{
CombineBaseCounts(x, na.rm = FALSE)
}
\arguments{
\item{x}{character vector in REDItools BaseCount format}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
A string with aggregated base counts
}
\description{
\code{CombineBaseCounts} sums a vector of base counts
}
\details{
The input to CombineBaseCounts should be a vector of character strings in
REDItools BaseCount format: [A, C, G, T] where A, C, G, and T represent the
number of reads with As, Cs, Gs, and Ts mapping to that site, respectively
}
\examples{
bc <- c("[6, 0, 4, 0]", "[2, 0, 8, 0]", NA)
CombineBaseCounts(bc, na.rm = TRUE)
}
