% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherMethodPValue.R
\name{FisherMethodPValue}
\alias{FisherMethodPValue}
\title{Perform Fisher's Method for combining p-values}
\usage{
FisherMethodPValue(pvals, zero.sub = 1e-05, na.rm = FALSE)
}
\arguments{
\item{pvals}{A matrix or data.frame containing the p-values from the single
tests}

\item{zero.sub}{Replacement for p-values of 0}

\item{na.rm}{Logical indicating whether NA values should be removed from the
analysis.}
}
\value{
A numeric vector with combined p-value
}
\description{
\code{FisherMethodPValue} combines p-values using Fisher's method.
}
\details{
Function for combining p-values by performing Fisher's method, which combines
p-values to a statistic:
\deqn{\chi_{2k}^k = -2 \sum_{i = 1}^{k} ln(p_i)}
}
\examples{
\dontrun{
pp <- matrix(c(runif(20),c(0.001,0.02,0.03,0.001)), ncol=4)
FisherMethodPValue(pp)
}
}
