% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAluRepeat.R
\name{GetAluRepeat}
\alias{GetAluRepeat}
\title{Categorize RNA Editing Events by Alu and Non-Alu Repeats}
\usage{
GetAluRepeat(x)
}
\arguments{
\item{x}{a vector giving UCSC repeat masking data}
}
\value{
A character vector with values "Alu", "Non-Alu Repeat", and
  "Non-Repetitive Element"
}
\description{
\code{GetAluRepeat} categorizes RNA editing events as falling within Alu
repeats, non-Alu repetitive elements, and non-repetitive regions
}
\examples{
\dontrun{
alu <- GetAluRepeat(RNAEdDataQCed$Annotation$RepMask_gid)
table(alu)
}
}
