% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGeneRegion.R
\name{GetGeneRegion}
\alias{GetGeneRegion}
\title{Categorize RNA Editing Events by Gene Region}
\usage{
GetGeneRegion(x)
}
\arguments{
\item{x}{a vector giving UCSC RefSeq annotation}
}
\value{
A character vector with values "3UTR", "5UTR", "CDS", "Exon",
  "Intron", and "Intergenic"
}
\description{
\code{GetGeneRegion} categorizes RNA editing events by genomic region
according to RefSeq annotation
}
\examples{
\dontrun{
regions <- GetGeneRegion(RNAEdDataQCed$Annotation$RefSeq_feat)
table(regions)
}
}
