% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSubType.R
\name{GetSubType}
\alias{GetSubType}
\title{Get Consensus Substitution Type for RNA Editing Event}
\usage{
GetSubType(x)
}
\arguments{
\item{x}{a matrix with substitution type observed at each editing event for
each individual (AllSubs)}
}
\value{
A character vector with most common substitution type for each
  editing event
}
\description{
\code{GetSubType} identifies the most common substitution type across samples
for each editing event
}
\examples{
\dontrun{
subtypes <- GetSubType(RNAEdDataQCed$AllSubs)
table(subtypes)
}
}
