% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Get ggplot default colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors to generate}
}
\value{
character vector specifying colors by hexadecimal code
}
\description{
\code{gg_color_hue} returns n colors using ggplot default palette
}
\examples{
\dontrun{
barplot(c(a = 2,b = 4, c= 3), col = gg_color_hue(3))
}
}
