% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalEditing.R
\name{GlobalEditing}
\alias{GlobalEditing}
\title{Get Global Editing Values}
\usage{
GlobalEditing(x, type = c("proportion", "count", "covered"), by = NULL)
}
\arguments{
\item{x}{An RNA editing data object which includes Annotation dataframe,
Frequency matrix, and Coverage-q25 matrix at minimum}

\item{type}{A character vector with values "proportion", "count", and/or
"covered" specifying global editing measure to return. By default, returns
all three.}

\item{by}{A character vector with name of column in x$Annotation dataframe by
which to split editing data}
}
\value{
A vector, matrix, or list of vectors or matrices with global editing
  values.
}
\description{
\code{GlobalEditing} returns global editing values for each sample.
}
\details{
The input should be a list object which includes matrices named "Frequency"
and "Coverage-q25". The function will return a vector, matrix, or list of
vectors or matrices with options to report the following:

\describe{
    \item{Proportion}{Total proportion of edited reads over reads mapping to edited sites (EditedReads / Coverage)}
    \item{Count}{Total number of sites with editing present}
    \item{Coverage}{Total number of sites with coverage (recommended for normalization of counts above)}
}

By default, calculates global values for all editing events combined. To
split by region, subtype, or other annotation, supply name of column in
Annotation data.frame to be used to split data.
}
\examples{
\dontrun{
GlobalEditing(RNAEdDataQCed)
GlobalEditing(RNAEdDataQCed, type = "proportion")
GlobalEditing(RNAEdDataQCed, type = "count", by = "SubType")
}
}
