% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyCorrectStrand.R
\name{IdentifyCorrectStrand}
\alias{IdentifyCorrectStrand}
\title{Identify Correct Strand for Editing Events}
\usage{
IdentifyCorrectStrand(annot, splitBy, values)
}
\arguments{
\item{annot}{Annotation dataframe}

\item{splitBy}{Character vector giving name of column by which to group
editing events}

\item{values}{Character vector with unique values to include}
}
\value{
A matrix with value name, correct strand for A, and correct strand
  for C.
}
\description{
\code{IdentifyCorrectStrand} is an internal function for identifying the
correct strand for editing events
}
