% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrandCorrection.R
\name{StrandCorrection}
\alias{StrandCorrection}
\title{Correct Strand Calling for RNA Editing Events}
\usage{
StrandCorrection(x, genes = NULL)
}
\arguments{
\item{x}{An RNA Editing data list object - first item of list must be
Annotation data.frame}

\item{genes}{A character string giving the name of the column in Annotation
data.frame with RefSeq gene annotations for gene-level strand correction}
}
\value{
An RNA Editing data object with strand correction
}
\description{
\code{StrandCorrection} identifies the correct strand for RNA editing calling
and corrects data
}
\examples{
\dontrun{
corrected <- StrandCorrection(RNAEdData, genes = "RefSeq_gid")
}
}
