% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrandCorrectionFuncs.R
\name{StrandCorrectionFuncs}
\alias{StrandCorrectionFuncs}
\title{Identify Functions to be Applied to RNA Editing Dataset for Strand Correction}
\usage{
StrandCorrectionFuncs(x)
}
\arguments{
\item{x}{An RNA Editing data list object - first item of list must be
Annotation data.frame}
}
\value{
A character vector with functions to be parsed in loop in
  StrandCorrection()
}
\description{
\code{StrandCorrectionFuncs} is an internal function that tests for the
presence of different matrix types in RNA editing data and lists appropriate
function to correct each
}
